+++ examples/analysis.py
@@ -17,10 +17,14 @@ import wandb
 import data
 import inept
 
+# Set env vars
+# os.environ['CUDA_VISIBLE_DEVICES']='0'
+# print(torch.cuda.device_count())
+
 # Get args
 import sys
-# run_id_idx = int(sys.argv[1])
-analysis_key_idx = int(sys.argv[1])
+run_id_idx = int(sys.argv[1])
+# analysis_key_idx = int(sys.argv[2])
 
 # Set params
 DEVICE = 'cuda:0' if torch.cuda.is_available() else 'cpu'
@@ -43,7 +47,6 @@ torch.set_grad_enabled(False);
 # - HIGH PRIORITY
 #   - Extent `get_present_func` to class with full state-altering capabilities
 #   - Add more accuracy metrics
-#   - Perturbation analysis with inverse transform
 #   - Add 2D functionality
 #   - Add optional UMAP
 # 
@@ -55,7 +58,9 @@ torch.set_grad_enabled(False);
 
 # %%
 # Parameters
+# run_id_idx = 1
 run_id = (
+    'brf6n6sn',  # TemporalBrain Random 100 Max
     'rypltvk5',  # MMD-MA Random 100 Max
     '32jqyk54',  # MERFISH Random 100 Max
     'c8zsunc9',  # ISS Random 100 Max
@@ -63,10 +68,10 @@ run_id = (
     'f6ajo2am',  # smFish NR
     'vb1x7bae',  # MERFISH NR
     '473vyon2',  # ISS NR
-)[0]
+)[run_id_idx]
 stage_override = None  # Manually override policy stage selection
 num_nodes_override = None
-max_batch_override = 1_000
+max_batch_override = None
 max_nodes_override = None
 seed_override = None
 
@@ -89,10 +94,11 @@ np.random.seed(seed)
 # Load data
 modalities, types, features = data.load_data(config['data']['dataset'], DATA_FOLDER)
 # config['data'] = inept.utilities.overwrite_dict(config['data'], {'standardize': True})  # Old model compatibility
+# config['data'] = inept.utilities.overwrite_dict(config['data'], {'top_variant': [20, 20]})  # Top variant testing
 if num_nodes_override is not None: config['data'] = inept.utilities.overwrite_dict(config['data'], {'num_nodes': num_nodes_override})
 if max_batch_override is not None: config['train'] = inept.utilities.overwrite_dict(config['train'], {'max_batch': max_batch_override})
 ppc = inept.utilities.Preprocessing(**config['data'], device=DEVICE)
-modalities = ppc.fit_transform(modalities)
+modalities, features = ppc.fit_transform(modalities, features)
 modalities, types = ppc.subsample(modalities, types)
 modalities = ppc.cast(modalities)
 
@@ -110,7 +116,8 @@ for file in run.files():
     if (stage_override is None and stage > latest_known_stage) or (stage_override is not None and stage == stage_override):
         if ftype == 'mdl': latest_mdl = [stage, file]
         elif ftype == 'wgt': latest_wgt = [stage, file]
-print(f'Policy found at stage {latest_mdl[0]}')
+print(f'MDL policy found at stage {latest_mdl[0]}')
+print(f'WGT policy found at stage {latest_wgt[0]}')
 
 # Load env
 env = inept.environments.trajectory(*modalities, **config['env'], **config['stages']['env'][0], device=DEVICE)
@@ -118,19 +125,19 @@ for weight_stage in config['stages']['env'][1:latest_mdl[0]+1]:
     env.set_rewards(weight_stage)
 
 # Load file
-load_type = 'mdl'
+load_type = 'wgt'
 if load_type == 'mdl':
     with tempfile.TemporaryDirectory() as tmpdir:
         latest_mdl[1].download(tmpdir, replace=True)
         policy = torch.load(os.path.join(tmpdir, latest_mdl[1].name))
 elif load_type == 'wgt':
-    # Mainly used in the case of old argument names, also more secure
+    # Mainly used in the case of old argument names, also generally more secure
     with tempfile.TemporaryDirectory() as tmpdir:
         latest_wgt[1].download(tmpdir, replace=True)
         # config['policy'] = inept.utilities.overwrite_dict(config['policy'], {'positional_dim': 6, 'modal_dims': [76]})  # Old model compatibility
         if max_nodes_override is not None: config['policy'] = inept.utilities.overwrite_dict(config['policy'], {'max_nodes': max_nodes_override})
         policy = inept.models.PPO(**config['policy'])
-        # incompatible_keys = policy.load_state_dict(torch.load(os.path.join(tmpdir, latest_wgt[1].name), weights_only=True))
+        incompatible_keys = policy.load_state_dict(torch.load(os.path.join(tmpdir, latest_wgt[1].name), weights_only=True))
 policy = policy.to(DEVICE).eval()
 policy.actor.set_action_std(1e-7)
 
@@ -144,20 +151,18 @@ times = types[0][:, -1]  # Temporary time annotation, will change per-dataset
 
 # %%
 # Choose key
-# TODO: Calculate both, plot one (?)
-analysis_key = [
-    'integration',
-    'discovery',
-    'temporal',
-][analysis_key_idx]
-state_manager_class = [
-    inept.utilities.IntegrationStateManager,
-    inept.utilities.DiscoveryStateManager,
-    inept.utilities.TemporalStateManager,
-][analysis_key_idx]
-
-# Discovery params
+# TODO: Calculate all, plot one (?)
+analysis_key_idx = 3
 discovery_key = 0  # Auto
+temporal_key = 0  # Auto
+perturbation_features = [np.random.choice(len(fs), 10, replace=False) for i, fs in enumerate(features) if (i not in env.reward_distance_target) or (len(env.reward_distance_target) == len(modalities))]
+
+analysis_key, state_manager_class = [
+    ('integration', inept.utilities.IntegrationStateManager),
+    ('discovery', inept.utilities.DiscoveryStateManager),
+    ('temporal', inept.utilities.TemporalStateManager),
+    ('perturbation', inept.utilities.PerturbationStateManager),
+][analysis_key_idx]
 
 # Discovery list
 discovery = []
@@ -172,9 +177,6 @@ discovery += [discovery_general]
 # Choose Discovery
 discovery = discovery[discovery_key]
 
-# Temporal params
-temporal_key = 0  # Auto
-
 # Stage order list
 temporal = []
 # Reverse alphabetical (ExSeq, MERFISH, smFISH, ISS, MouseVisual)
@@ -183,11 +185,11 @@ temporal += [temporal_general]
 # Choose stage order
 temporal = temporal[temporal_key]
 
+# Perturbation feature names
+perturbation_feature_names = [[fnames[pf] for pf in pfs] for pfs, fnames in zip(perturbation_features, features)]
+
 # Initialize memories
 memories = {}
-
-# %% [markdown]
-# ##### Main Run
 
 # %%
 # Choose state manager
@@ -195,8 +197,12 @@ state_manager = state_manager_class(
     device=DEVICE,
     discovery=discovery,
     temporal=temporal,
+    perturbation_features=perturbation_features,
+    modal_targets=env.reward_distance_target,
     num_nodes=modalities[0].shape[0],
+    dim=env.dim,
 )
+get_current_stage = lambda: state_manager.current_stage if state_manager_class in (inept.utilities.TemporalStateManager, inept.utilities.PerturbationStateManager) else -1
 
 # Initialize
 env.reset(); memories[analysis_key] = defaultdict(lambda: [])
@@ -205,17 +211,19 @@ env.reset(); memories[analysis_key] = defaultdict(lambda: [])
 state_vars, end = state_manager(
     # present=present,
     state=env.get_state(),
+    modalities=ppc.cast(ppc.inverse_transform(ppc.inverse_cast(env.get_modalities()))),
     labels=labels,
     times=times,
 )
 present = state_vars['present']
 env.set_state(state_vars['state'])
+raw_modalities = state_vars['modalities']
+env.set_modalities(ppc.cast(ppc.transform(ppc.inverse_cast(raw_modalities))))
 
 # Continue initializing
 memories[analysis_key]['present'].append(present)
 memories[analysis_key]['states'].append(env.get_state())
-current_stage = state_manager.current_stage if state_manager_class in (inept.utilities.TemporalStateManager,) else -1
-memories[analysis_key]['stages'].append(current_stage)
+memories[analysis_key]['stages'].append(get_current_stage())
 memories[analysis_key]['rewards'].append(torch.zeros(modalities[0].shape[0], device=DEVICE))
 
 # Simulate
@@ -224,7 +232,7 @@ while True:
     # CLI
     if timestep % 20 == 0:
         cli_out = f'Timestep: {timestep}'
-        if analysis_key == 'temporal': cli_out += f' - Stage: {current_stage}'
+        if get_current_stage() != -1: cli_out += f' - Stage: {get_current_stage()}'
         print(cli_out, end='\r')
 
     # Step
@@ -246,17 +254,21 @@ while True:
     state_vars, end = state_manager(
         present=present,
         state=env.get_state(),
+        modalities=raw_modalities,
         labels=labels,
         times=times,
     )
     present = state_vars['present']
     env.set_state(state_vars['state'])
+    # Only modify if changes
+    if torch.tensor([(rm != svm).any() for rm, svm in zip(raw_modalities, state_vars['modalities'])]).any():
+        raw_modalities = state_vars['modalities']
+        env.set_modalities(ppc.cast(ppc.transform(ppc.inverse_cast(raw_modalities))))
 
     # Record
     memories[analysis_key]['present'].append(present)
     memories[analysis_key]['states'].append(env.get_state())
-    current_stage = state_manager.current_stage if state_manager_class in (inept.utilities.TemporalStateManager,) else -1
-    memories[analysis_key]['stages'].append(current_stage)
+    memories[analysis_key]['stages'].append(get_current_stage())
     memories[analysis_key]['rewards'].append(rewards)
 
     # End
@@ -274,7 +286,52 @@ memories[analysis_key] = dict(memories[analysis_key])
 print()
 
 # %%
+stages, counts = np.unique(memories[analysis_key]['stages'], return_counts=True)
+print('Steps per Stage:')
+for s, c in zip(stages, counts):
+    print(f'{s}: {c}')
+
+# %%
 memories[analysis_key]['rewards'].cpu().mean()  # TODO: Why not the same as WANDB?
+
+# %%
+# Perturbation significance analysis
+if 'perturbation' in memories:
+    # Get last idx for each stage
+    stages = memories['perturbation']['stages'].cpu().numpy()
+    unique_stages, unique_idx = np.unique(stages[::-1], return_index=True)
+    unique_idx = stages.shape[0] - unique_idx - 1
+    # unique_stages, unique_idx = unique_stages[::-1], unique_idx[::-1]
+
+    # Record perturbation feature pairs
+    perturbation_feature_triples = [(i, f, n) for i, (fs, ns) in enumerate(zip(perturbation_features, perturbation_feature_names)) for f, n in zip(fs, ns)]
+
+    # Compute effect sizes for each
+    effect_sizes = []
+    for stage, idx in zip(unique_stages, unique_idx):
+        # Get state
+        state = memories['perturbation']['states'][idx]
+
+        # Record steady state after integration
+        if stage == 0:
+            steady_state = state
+            continue
+
+        # Get perturbed feature
+        m_idx, pf, pf_name = perturbation_feature_triples[stage-1]
+
+        # Compute effect size
+        effect_size = (state[:, :env.dim] - steady_state[:, :env.dim]).square().sum(dim=-1).sqrt().mean(dim=-1).item()
+        effect_sizes.append(effect_size)
+
+    # Print effect sizes
+    i = 0
+    for j, (pfs, pfns) in enumerate(zip(perturbation_features, perturbation_feature_names)):
+        print(f'Modality {j}:')
+        for pf, pfn in zip(pfs, pfns):
+            print(f'{pfn}:\t{effect_sizes[i]:.02e}')
+            i += 1
+        print()
 
 # %% [markdown]
 # ### Plot Memories
@@ -293,13 +350,13 @@ modal_dist = env.dist
 
 # Parameters
 interval = 1e3*env.delta/3  # Time between frames (3x speedup)
-min_max_vel = -1 if analysis_key == 'temporal' else 1e-2  # Stop at first frame all vels are below target. 0 for full play
+min_max_vel = 1e-2 if analysis_key in ('integration', 'discovery') else -1  # Stop at first frame all vels are below target. 0 for full play
 frame_override = None  # Manually enter number of frames to draw
 num_lines = 25  # Number of attraction and repulsion lines
 rotations_per_second = .1  # Camera azimuthal rotations per second
 
 # Create plot based on key
-# NOTE: Standard 1-padding all around, then 3 between figures
+# NOTE: Standard 1-padding all around and between figures
 # NOTE: Left, bottom, width, height
 if analysis_key in ('integration', 'discovery'):
     figsize = (15, 10)
@@ -329,6 +386,22 @@ elif analysis_key == 'temporal':
         inept.utilities.ViewTemporalDiscrepancy,
     ]
 
+elif analysis_key in ('perturbation',):
+    figsize = (20, 10)
+    fig = plt.figure(figsize=figsize)
+    axs = [
+        fig.add_axes([1 /figsize[0], 1 /figsize[1], 8 /figsize[0], 8 /figsize[1]], projection='3d'),
+        fig.add_axes([10 /figsize[0], 5.5 /figsize[1], 8 /figsize[0], 3.5 /figsize[1]]),
+        fig.add_axes([10 /figsize[0], 1 /figsize[1], 3.5 /figsize[0], 3.5 /figsize[1]]),
+        fig.add_axes([14.5 /figsize[0], 1 /figsize[1], 3.5 /figsize[0], 3.5 /figsize[1]]),
+    ]
+    views = [
+        inept.utilities.View3D,
+        inept.utilities.ViewPerturbationEffect,
+        inept.utilities.ViewTemporalScatter,
+        inept.utilities.ViewSilhouette,
+    ]
+
 # Initialize views
 arguments = {
     # Data
@@ -342,6 +415,8 @@ arguments = {
     'dim': env.dim,
     'modal_targets': env.reward_distance_target,
     'temporal_stages': temporal['stages'] if analysis_key == 'temporal' else None,
+    'perturbation_features': perturbation_features,
+    'perturbation_feature_names': perturbation_feature_names,
     # Arguments
     'interval': interval,
     'skip': skip,